backend/requirements.txt
django==3.2
djangorestframework
django-cors-headers
cloudinary
djangorestframework-simplejwt
mysqlclient
Pillow
python-decouple